Unified button shadow & InfoDialog changes
Makes the ! button larger and more in-line with the other buttons on the page.  The wording of the infobox has been slightly adjusted

Changes all buttons sitewide to a 1px shadow blur, all the buttons on the map now use the same blur / shadow levels.

Map search now reads "search for location" in order to further the distinction between the add a node workflow (which needs to be changed in the future but that's a future thing).

Changes the sliders and removes the dumb labels.... I hate those labels :P  They're also snappier now.